fixing get_fps()
git-svn-id: svn://dyne.org/rastasoft/freej/branches/Goil/freej_91@1080 383723c8-4afa-0310-b8a8-b1afb83214fc

+++ src/jsync.cpp
@@ -132,7 +132,7 @@ void JSyncThread::calc_fps() {
 	fpsd.i %= fpsd.n;
 }
 
-float JSyncThread::get_fps() {
+double JSyncThread::get_fps() {
 	return (fps ? fpsd.sum / fpsd.n : 0 );
 }
